libmbim-glib,proxy: fix reporting of merged subscribe list
We were doing the subscribe list merge correctly, but not reporting it
as done because the 'updated' pointer is set to NULL before returning
it.

This was breaking QMI over MBIM indications, as the registration
message to request them was never sent to the device.

(cherry picked from commit ecd188ff0126866199fd453e962ba54ba1dce290)