meson.build: add an option to not build man pages.
When cross-compiling, the compiled binaries cannot be run on the build
machine, unless one defines a helper (e.g. qemu) in the meson
configuration, which is most of the time undesirable.

If help2man is installed on the build machine, then the build would fail
because of this, so add an option to disable generating the man pages.

With later versions of meson, it will be possible to automatically
detect this condition.

(cherry picked from commit 792af73ea90e9689703ab526dd444fdb1c2f8e40)